tests/wpt/tests/svg/types/scripted/SVGGraphicsElement.getScreenCTM.html

<!DOCTYPE html>
<html>
<head>
  <title>SVGGraphicsElement.getScreenCTM</title>
  <metadata>
    <link rel="help" href="https://svgwg.org/svg2-draft/types.html#InterfaceSVGGraphicsElement"/>
  </metadata>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<style>
  body {
    margin: 8px;
  }
  #container1 {
    transform: translate(50px, 50px);
  }
  #container2 {
    position: absolute;
    top: 0;
    left: 0;
    transform: scale(2, 1);
  }
</style>
<div id="container1">
  <svg id="svg1" width="150" height="150">
    <rect id="rect1" fill="lime" x="50" y="50" width="100" height="100"/>
  </svg>
</div>
<div id="container2">
  <svg id="svg2" width="100" height="100"/>
</div>
<script>
test(function() {
  let pt = DOMPoint.fromPoint({x: 58, y: 58});
  let screenCTM = document.getElementById("svg1").getScreenCTM();
  assert_equals(screenCTM.a, 1);
  assert_equals(screenCTM.d, 1);
  assert_equals(screenCTM.e, 58);
  assert_equals(screenCTM.f, 58);

  let transformedPoint = pt.matrixTransform(screenCTM.inverse());

  assert_equals(transformedPoint.x, 0);
  assert_equals(transformedPoint.y, 0);
});

test(function() {
  let screenCTM = document.getElementById("rect1").getScreenCTM();
  assert_equals(screenCTM.a, 1);
  assert_equals(screenCTM.d, 1);
  assert_equals(screenCTM.e, 58);
  assert_equals(screenCTM.f, 58);
});

test(function() {
  let screenCTM = document.getElementById("svg2").getScreenCTM();
  assert_equals(screenCTM.a, 2);
  assert_equals(screenCTM.d, 1);
});

</script>
</body>
</html>
